Added documentation, nerfed Pin Missile back down, and adjusted Rock Slide.
Added a changelog and a detailed explanation of the changes to be transparent on the purposes they serve.  Will also make it easier for me to make official documentation in the future if that ever becomes necessary.
 I also cleaned up the README and removed the FAQ becuase I'll just point people to the Pokémon Crystal disassembly repo instead.

I lowered Pin Missile back down to 15 base power but upped its accuracy to 100.  Makes it more fitting for an early game move now.

I further increased Rock Slide's power to 100 to better differentiate it from Strength and lowered its flinch chance and PP to compensate.