Update Beat Up
Changed it from a multi-hit move to a single-hit move with a chance to lower defense and made it a custom physical attack.  80 power, 100 accuracy, 15 PP, 20% chance to lower defense.

Removed several instances of Beat Up's vanilla code such as unique animation commands.

Removed Will-O-Wisp from Vulpix's level-up (forgot to remove it as part of testing).

Also lowered Crunch's proc rate for SpD lowering to 10% to compensate for boosted damage.